hash: Make basis of hash_words64() 32 bits.
The basis of hash_words64() was 64 bits, even when the hash value is
32 bits, thus confusing the domain and the range of the function.
This patch fixes this by making the basis an uint32_t.

Suggested-by: Ben Pfaff <blp@nicira.com>

+++ lib/hash.c
@@ -67,7 +67,7 @@ hash_words__(const uint32_t p[], size_t n_words, uint32_t basis)
 }
 
 uint32_t
-hash_words64__(const uint64_t p[], size_t n_words, uint64_t basis)
+hash_words64__(const uint64_t p[], size_t n_words, uint32_t basis)
 {
     return hash_words64_inline(p, n_words, basis);
 }
